move accessor type
